// samples/web-csr/dressca-frontend/admin/src/generated/api-client/models/catalog-brand-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Dressca Web API
 * Dressca の Web API 仕様
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * カタログブランドのレスポンスデータを表します。 カタログアイテムの製造元や企画元に基づいて定義されるブランドを表現します。             
 * @export
 * @interface CatalogBrandResponse
 */
export interface CatalogBrandResponse {
    /**
     * カタログブランド Id を取得または設定します。             
     * @type {number}
     * @memberof CatalogBrandResponse
     */
    'id': number;
    /**
     * ブランド名を取得または設定します。             
     * @type {string}
     * @memberof CatalogBrandResponse
     */
    'name': string;
}