// samples/web-csr/dressca-frontend/admin/src/generated/api-client/models/catalog-category-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Dressca Web API
 * Dressca の Web API 仕様
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * カタログカテゴリのレスポンスデータを表します。             
 * @export
 * @interface CatalogCategoryResponse
 */
export interface CatalogCategoryResponse {
    /**
     * カタログカテゴリ Id を取得または設定します。             
     * @type {number}
     * @memberof CatalogCategoryResponse
     */
    'id': number;
    /**
     * カテゴリ名を取得または設定します。             
     * @type {string}
     * @memberof CatalogCategoryResponse
     */
    'name': string;
}